Initialize properties in add_person (#79)
Adds the ability to initialize properties in add_person(), as in:

       let person = context.add_person(((Age, 17), (IsSwimmer, true))).unwrap()

Any non-derived property that doesn't have automatic initialization must be initialized in this fashion during creation, which means we have removed inititialize_person_property(). This is technically enforced by checking that any known such property is present in the initialization list. This enforcement is imperfect because Context doesn't know about any property which has never been used. This means that if you do add_person() followed by get_person_property() for such a property, it will panic. However, as long as there is any use of a property this invariant will instead be enforced in add_person()