Initial implementation of a network facility.
A network is modeled as a directed graph.  Edges are typed in the
usual fashion, i.e., keyed by a Rust type, and each person can have an
arbitrary number of outgoing edges of a given type, with each edge
having a weight, as in:

        context
            .add_edge::<EdgeType1>(person1, person2, 0.01);
        context
            .add_edge::<EdgeType1>(person1, person3, 0.02);

Which creates a graph where `person1` has two outgoing edges of type
`EdgeType`, to `person2` and `person3`.

Edges are directional but `add_edge_bidi()` will create one edge in
each direction between two people with the same weight for both
edges. If you want to create edges with different weights, you have to
do it separately.

Creating an edge to yourself is forbidden (returns IxaError). It is
not technically possible to create an edge to a person who doesn't
exist because `PersonId` is minted by `add_person()`.